% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhat-package.R
\docType{package}
\name{Bhat-package}
\alias{Bhat}
\alias{Bhat-package}
\title{Bhat: General likelihood exploration}
\description{
This package provides functions for Maximum Likelihood Estimation, Markov Chain Monte Carlo, finding confidence intervals. This implementation is heavily based on the original Fortran source code translated to R.
}
\author{
\strong{Maintainer}: Rafael Meza \email{rmeza@umich.edu} (\href{https://orcid.org/0000-0002-1076-5037}{ORCID})

Authors:
\itemize{
  \item Georg Luebeck \email{gluebeck@fhcrc.org} (\href{https://orcid.org/0000-0002-4378-2829}{ORCID})
}

Other contributors:
\itemize{
  \item Alexander Gaenko \email{galexv@umich.edu} (\href{https://orcid.org/0000-0002-1901-0394}{ORCID}) [contributor]
}

}
\keyword{internal}
