% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp.R
\name{dfp}
\alias{dfp}
\title{Function minimization with box-constraints}
\usage{
dfp(x, f, tol = 1e-05, nfcn = 0, ...)
}
\arguments{
\item{x}{a list with components 'label' (of mode character), 'est' (the
parameter vector with the initial guess), 'low' (vector with lower bounds),
and 'upp' (vector with upper bounds)}

\item{f}{the function that is to be minimized over the parameter vector
defined by the list \code{x}}

\item{tol}{a tolerance used to determine when convergence should be
indicated}

\item{nfcn}{number of function calls}

\item{...}{other parameters to be passed to `f`}
}
\value{
list with the following components: \item{fmin }{ the function
value f at the minimum } \item{label }{ the labels taken from list \code{x}
} \item{est }{ a vector of the estimates at the minimum. dfp does not
overwrite \code{x} } \item{status }{ 0 indicates convergence, 1 indicates
non-convergence } \item{nfcn }{ no. of function calls }
}
\description{
This Davidon-Fletcher-Powell optimization algorithm has been `hand-tuned'
for minimal setup configuration and for efficency. It uses an internal
logit-type transformation based on the pre-specified box-constraints.
Therefore, it usually does not require rescaling (see help for the R optim
function). \code{dfp} automatically computes step sizes for each parameter
to operate with sufficient sensitivity in the functional output.
Performance is comparable to the BFGS algorithm in the R function
\code{optim}. \code{dfp} interfaces with \code{newton} to ascertain
convergence, compute the eigenvalues of the Hessian, and provide 95\%
confidence intervals when the function to be minimized is a negative
log-likelihood.
}
\details{
The dfp function minimizes a function \code{f} over the parameters
specified in the input list \code{x}. The algorithm is based on Fletcher's
"Switching Method" (Comp.J. 13,317 (1970))

the code has been 'transcribed' from Fortran source code into R
}
\note{
This function is part of the Bhat exploration tool
}
\examples{

        # generate some Poisson counts on the fly
          dose <- c(rep(0,50),rep(1,50),rep(5,50),rep(10,50))
          data <- cbind(dose,rpois(200,20*(1+dose*.5*(1-dose*0.05))))

        # neg. log-likelihood of Poisson model with 'linear-quadratic' mean: 
          lkh <- function (x) { 
          ds <- data[, 1]
          y  <- data[, 2]
          g <- x[1] * (1 + ds * x[2] * (1 - x[3] * ds)) 
          return(sum(g - y * log(g)))
          }

	# for example define
          x <- list(label=c("a","b","c"),est=c(10.,10.,.01),low=c(0,0,0),upp=c(100,20,.1))

	# call:
	  results <- dfp(x,f=lkh)

}
\references{
Fletcher's Switching Method (Comp.J. 13,317, 1970)
}
\seealso{
optim, \code{\link{newton}}, \code{\link{ftrf}},
\code{\link{btrf}}, \code{\link{logit.hessian}}
}
\author{
E. Georg Luebeck (FHCRC)
}
\keyword{methods}
\keyword{optimize}
